[fix]jsonhelperをcommunicationmanager内に置いた
