+++ Assets/Script.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 63e21c51ae35812429d72c3bbf82bde5
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
