+++ Assets/Script/CommunicationManager.cs.meta
